prepared BackupSession and RestoreSession common code
