+++ controllers/backupsession_controller_helpers.go
@@ -12,49 +12,6 @@ import (
 	"strings"
 )
 
-const (
-	RESTIC_EXEC = "/usr/bin/restic"
-)
-
-func (r *BackupSessionReconciler) setResticEnv(backupConf formolv1alpha1.BackupConfiguration) error {
-	repo := formolv1alpha1.Repo{}
-	if err := r.Get(r.Context, client.ObjectKey{
-		Namespace: backupConf.Namespace,
-		Name:      backupConf.Spec.Repository,
-	}, &repo); err != nil {
-		r.Log.Error(err, "unable to get repo")
-		return err
-	}
-	if repo.Spec.Backend.S3 != nil {
-		os.Setenv(formolv1alpha1.RESTIC_REPOSITORY, fmt.Sprintf("s3:http://%s/%s/%s-%s",
-			repo.Spec.Backend.S3.Server,
-			repo.Spec.Backend.S3.Bucket,
-			strings.ToUpper(backupConf.Namespace),
-			strings.ToLower(backupConf.Name)))
-		data := r.getSecretData(repo.Spec.RepositorySecrets)
-		os.Setenv(formolv1alpha1.AWS_SECRET_ACCESS_KEY, string(data[formolv1alpha1.AWS_SECRET_ACCESS_KEY]))
-		os.Setenv(formolv1alpha1.AWS_ACCESS_KEY_ID, string(data[formolv1alpha1.AWS_ACCESS_KEY_ID]))
-		os.Setenv(formolv1alpha1.RESTIC_PASSWORD, string(data[formolv1alpha1.RESTIC_PASSWORD]))
-	}
-	return nil
-}
-
-func (r *BackupSessionReconciler) checkRepo() error {
-	r.Log.V(0).Info("Checking repo")
-	if err := exec.Command(RESTIC_EXEC, "unlock").Run(); err != nil {
-		r.Log.Error(err, "unable to unlock repo", "repo", os.Getenv(formolv1alpha1.RESTIC_REPOSITORY))
-	}
-	output, err := exec.Command(RESTIC_EXEC, "check").CombinedOutput()
-	if err != nil {
-		r.Log.V(0).Info("Initializing new repo")
-		output, err = exec.Command(RESTIC_EXEC, "init").CombinedOutput()
-		if err != nil {
-			r.Log.Error(err, "something went wrong during repo init", "output", output)
-		}
-	}
-	return err
-}
-
 type BackupResult struct {
 	SnapshotId string
 	Duration   float64
