+++ controllers/server.go
@@ -37,6 +37,16 @@ func StartServer() {
 		setupLog.Error(err, "unable to create manager")
 		os.Exit(1)
 	}
+	if err = (&RestoreSessionReconciler{
+		Session: Session{
+			Client: mgr.GetClient(),
+			Scheme: mgr.GetScheme(),
+		},
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "RestoreSession")
+		os.Exit(1)
+	}
+
 	if err = (&BackupSessionReconciler{
 		Session: Session{
 			Client: mgr.GetClient(),
